Return all available connection information on each RDS reconcile
This updates the RDS controller to return any available
connection details on each reconcile. Some are expected
to only be available in Create or Observe, but using a
consistent strategy for propagating makes reasoning about
which values are being updated simpler and avoids having
to specify connection details in each location.

Signed-off-by: hasheddan <georgedanielmangum@gmail.com>

$Signed-off-by: hasheddan <georgedanielmangum@gmail.com>